refactor waypoint listeners to new file
